# docs/openapi-consumer.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "X-Ray 企业变化解释器",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/v2/discovery": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Discovery",
        "description": "Network discovery is opt-in; no request writes verified company evidence.",
        "operationId": "discovery_api_v2_discovery_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 80,
              "title": "Query"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/consumer/qcc-session": {
      "get": {
        "tags": [
          "消费者实时调查"
        ],
        "summary": "Qcc Session Status",
        "operationId": "qcc_session_status_api_consumer_qcc_session_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "消费者实时调查"
        ],
        "summary": "Set Qcc Session",
        "operationId": "set_qcc_session_api_consumer_qcc_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/consumer/workbuddy/config": {
      "post": {
        "tags": [
          "消费者实时调查"
        ],
        "summary": "Configure Workbuddy",
        "operationId": "configure_workbuddy_api_consumer_workbuddy_config_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkBuddyConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/consumer/workbuddy/callback": {
      "get": {
        "tags": [
          "消费者实时调查"
        ],
        "summary": "Workbuddy Callback",
        "operationId": "workbuddy_callback_api_consumer_workbuddy_callback_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "State"
            }
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Error"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/consumer/discovery": {
      "post": {
        "tags": [
          "消费者实时调查"
        ],
        "summary": "Discover",
        "operationId": "discover_api_consumer_discovery_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__schemas__consumer__DiscoveryInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Discovery"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/consumer/analyses": {
      "post": {
        "tags": [
          "消费者实时调查"
        ],
        "summary": "Analyse",
        "operationId": "analyse_api_consumer_analyses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Analysis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/consumer/jobs": {
      "post": {
        "tags": [
          "消费者实时调查"
        ],
        "summary": "Start Job",
        "operationId": "start_job_api_consumer_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/consumer/jobs/{ident}": {
      "get": {
        "tags": [
          "消费者实时调查"
        ],
        "summary": "Get Job",
        "operationId": "get_job_api_consumer_jobs__ident__get",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "消费者实时调查"
        ],
        "summary": "Cancel Job",
        "operationId": "cancel_job_api_consumer_jobs__ident__delete",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Analysis": {
        "properties": {
          "analysis_id": {
            "type": "string",
            "title": "Analysis Id"
          },
          "company_id": {
            "type": "string",
            "title": "Company Id"
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At"
          },
          "evidence_as_of": {
            "type": "string",
            "title": "Evidence As Of"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "fallback": {
            "type": "boolean",
            "title": "Fallback"
          },
          "coverage_status": {
            "type": "string",
            "title": "Coverage Status"
          },
          "identity": {
            "$ref": "#/components/schemas/IdentityCandidate"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "changes": {
            "items": {
              "$ref": "#/components/schemas/Change"
            },
            "type": "array",
            "title": "Changes"
          },
          "questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Questions"
          },
          "unknowns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unknowns"
          },
          "comparables": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Comparables"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Sources"
          },
          "trace": {
            "items": {
              "$ref": "#/components/schemas/Step"
            },
            "type": "array",
            "title": "Trace"
          },
          "criteria": {
            "additionalProperties": true,
            "type": "object",
            "title": "Criteria"
          },
          "counter_search_status": {
            "type": "string",
            "title": "Counter Search Status"
          },
          "counter_source_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Counter Source Ids"
          },
          "agent_status": {
            "type": "string",
            "title": "Agent Status"
          },
          "indicators": {
            "items": {
              "$ref": "#/components/schemas/Indicator"
            },
            "type": "array",
            "title": "Indicators"
          },
          "agent_framework": {
            "type": "string",
            "title": "Agent Framework",
            "default": "LangGraph"
          },
          "agent_model_used": {
            "type": "boolean",
            "title": "Agent Model Used",
            "default": false
          },
          "agent_rounds": {
            "type": "integer",
            "title": "Agent Rounds",
            "default": 0
          },
          "risk": {
            "$ref": "#/components/schemas/RiskAssessment"
          },
          "source_stats": {
            "additionalProperties": true,
            "type": "object",
            "title": "Source Stats"
          },
          "reviews": {
            "$ref": "#/components/schemas/ReviewAssessment"
          },
          "cashflow": {
            "$ref": "#/components/schemas/CashflowAssessment"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "analysis_id",
          "company_id",
          "generated_at",
          "evidence_as_of",
          "mode",
          "fallback",
          "coverage_status",
          "identity",
          "summary",
          "changes",
          "questions",
          "unknowns",
          "sources",
          "trace",
          "criteria",
          "counter_search_status",
          "counter_source_ids",
          "agent_status",
          "indicators"
        ],
        "title": "Analysis"
      },
      "AnalysisInput": {
        "properties": {
          "investigation_id": {
            "type": "string",
            "maxLength": 64,
            "title": "Investigation Id"
          },
          "candidate_id": {
            "type": "string",
            "maxLength": 64,
            "title": "Candidate Id"
          },
          "intent": {
            "type": "string",
            "enum": [
              "initial_purchase",
              "top_up",
              "renewal",
              "explore"
            ],
            "title": "Intent",
            "default": "explore"
          },
          "service_category": {
            "type": "string",
            "enum": [
              "fitness",
              "education",
              "beauty",
              "eldercare",
              "other"
            ],
            "title": "Service Category",
            "default": "fitness"
          },
          "amount_yuan": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Yuan"
          },
          "service_duration_months": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 120.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Duration Months"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "investigation_id",
          "candidate_id"
        ],
        "title": "AnalysisInput"
      },
...
      "Discovery": {
        "properties": {
          "investigation_id": {
            "type": "string",
            "title": "Investigation Id"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At"
          },
          "mode": {
            "type": "string",
            "enum": [
              "live_search",
              "cached",
              "unavailable"
            ],
            "title": "Mode"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/IdentityCandidate"
            },
            "type": "array",
            "title": "Candidates"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Sources"
          },
          "trace": {
            "items": {
              "$ref": "#/components/schemas/Step"
            },
            "type": "array",
            "title": "Trace"
          },
          "unknowns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unknowns"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "investigation_id",
          "query",
          "location",
          "generated_at",
          "mode",
          "candidates",
          "sources",
          "trace",
          "unknowns"
        ],
        "title": "Discovery"
      },
      "DiscoveryResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "companies": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Companies"
          },
          "announcements": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Announcements"
          },
          "verification_status": {
            "type": "string",
            "title": "Verification Status"
          },
          "external_search_links": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "External Search Links"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "status",
          "query",
          "companies",
          "announcements",
          "verification_status",
          "external_search_links",
          "limitations"
        ],
        "title": "DiscoveryResponse"
      },
...
      "Evidence": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "publisher": {
            "type": "string",
            "title": "Publisher"
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "date_semantics": {
            "type": "string",
            "title": "Date Semantics",
            "default": "公开日期未核实"
          },
          "fetched_at": {
            "type": "string",
            "title": "Fetched At"
          },
          "verification_status": {
            "type": "string",
            "enum": [
              "search_excerpt",
              "page_text",
              "provider_response"
            ],
            "title": "Verification Status",
            "default": "search_excerpt"
          },
          "channel": {
            "type": "string",
            "enum": [
              "public_web",
              "community",
              "government",
              "registry"
            ],
            "title": "Channel",
            "default": "public_web"
          },
          "purpose": {
            "type": "string",
            "title": "Purpose"
          },
          "page_status": {
            "type": "string",
            "title": "Page Status",
            "default": "未读取原网页"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "default": false
          },
          "scope": {
            "type": "string",
            "enum": [
              "selected_entity",
              "brand_context"
            ],
            "title": "Scope",
            "default": "selected_entity"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "url",
          "publisher",
          "excerpt",
          "fetched_at",
          "purpose"
        ],
        "title": "Evidence"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IdentityCandidate": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "basis": {
            "type": "string",
            "title": "Basis"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Ids"
          },
          "relationship_status": {
            "type": "string",
            "title": "Relationship Status",
            "default": "网页提及，门店关联尚未核实"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "basis",
          "source_ids"
        ],
        "title": "IdentityCandidate"
      },
...
      "SessionInput": {
        "properties": {
          "cookie": {
            "type": "string",
            "format": "password",
            "title": "Cookie",
            "writeOnly": true
          }
        },
        "type": "object",
        "required": [
          "cookie"
        ],
        "title": "SessionInput"
      },
...
      "Step": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "action",
          "status",
          "detail"
        ],
        "title": "Step"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkBuddyConfig": {
        "properties": {
          "client_id": {
            "type": "string",
            "maxLength": 256,
            "minLength": 2,
            "title": "Client Id"
          },
          "client_secret": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 2,
            "format": "password",
            "title": "Client Secret",
            "writeOnly": true
          },
          "redirect_uri": {
            "type": "string",
            "maxLength": 2048,
            "title": "Redirect Uri"
          }
        },
        "type": "object",
        "required": [
          "client_id",
          "client_secret",
          "redirect_uri"
        ],
        "title": "WorkBuddyConfig"
      },
...
      "app__schemas__consumer__DiscoveryInput": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 80,
            "minLength": 2,
            "pattern": "^[^\\x00-\\x1f\\x7f]+$",
            "title": "Query"
          },
          "location": {
            "type": "string",
            "maxLength": 60,
            "pattern": "^[^\\x00-\\x1f\\x7f]*$",
            "title": "Location",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query"
        ],
        "title": "DiscoveryInput"
      }
    }
  }
}